capture avoidance principle reframed as "expanded form" principle
As written, the capture avoidance principle was the one that actually asserts
the `(eexpanded : tsplice -> texpansion) (esplice)` form which all expanded
results will inevitably take on. Since this form, or a simplification of it,
was a premise of some of the other reasoning principles, it made sense to
reframe the "capture avoidance" principle as the first principle and have its
purpose be to assert this form. I suppose there may still be a need for a
separate "capture avoidance" principle, stated right before the
"context independence" principle, but i don't know what is the best way to
express it. We can discuss the best way to formally express the conceptual
principle of "capture avoidance", and then add that back in, but either way
I think it is good to have the "expanded form" principle be a distinct
principle that is stated before the others, so that the expression of the
other principles is clear and unpresumptuous.