Revert last 3 commits, for new proj-stepping approach
The previous 3 commits explored an approach in which there is a step
rule that moves a cast on a pair inside the pair, but to maintain
consistency with prior work, we instead take a "proj-stepping" approach,
in which we have rules that will help a projection step even if its
argument has a cast.
This revert commit seems the simplest way to preserve the prior approach
in this branch.