working on #2.
i've switched from building the very latest version from source to pulling
a binary from apt into an ubuntu image. this takes a lot less time and is
far less likely to fail; the source build process is quite brittle. this
means we're a build version behind but i don't think that matters too
much. i'll note that in the readme.

i'm currently not sure where the output from actually running agda goes
once it's built; i can log into the container and see that agda is
installed and the source is copied over but the command i think should run
it terminates suspiciously quickly and silently.