close #2
updated the readme and remembered that you have to actually run the
container to get it to run.

there is a bizzare issue with agda where if you run it on `-v 2` to produce
even more output, it crashes as soon as it sees a unicode character. this
is known and documented here but not fixed in the version on apt:
https://github.com/agda/agda/issues/2922

i have worked around this by running it on a lower verbosity, which is
closer to what we actually want anyway. the downside is that the happy
condition for exit is that it doesn't emit an error code, it just falls off
the end of `all.agda`. it used to print "finished" or similar. that's fine,
i think, and i updated the README. the next step up in verbostiy both
crashes on unicode and prints out way more fiddly detail than we'd want an
AE reviewer to see anyway.